# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a API documentation for OB server.  You can find out more about Ocean Builders at [https://ocean.builders/](https://ocean.builders/)",
    "version": "0.7",
    "title": "Ocean Builder",
    "contact": {
      "email": "diego@scissorboy.com"
    }
  },
  "host": "oceanbuilders.herokuapp.com",
  "basePath": "/v1/api/",
...
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/seapods": {
      "post": {
        "tags": [
          "Seapods"
        ],
        "summary": "Create New Seapod",
        "description": "",
        "operationId": "createSeapod",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Seapod information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "SeaPodName": {
                  "type": "string"
                },
                "exteriorFinish": {
                  "type": "string"
                },
                "exterioirColor": {
                  "type": "string",
                  "pattern": "^0xFF[A-F0-9]{6}$"
                },
                "sparFinish": {
                  "type": "string"
                },
                "sparDesign": {
                  "type": "string"
                },
                "deckEnclosure": {
                  "type": "string"
                },
                "bedAndLivingRoomEnclousure": {
                  "type": "string"
                },
                "power": {
                  "type": "string"
                },
                "powerUtilities": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "underWaterRoomFinishing": {
                  "type": "string"
                },
                "underWaterWindows": {
                  "type": "string"
                },
                "soundSystem": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "masterBedroomFloorFinishing": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "masterBedroomInteriorWallColor": {
                  "type": "string",
                  "pattern": "^0xFF[A-F0-9]{6}$"
                },
                "livingRoomloorFinishing": {
                  "type": "string"
                },
                "livingRoomInteriorWallColor": {
                  "type": "string",
                  "pattern": "^0xFF[A-F0-9]{6}$"
                },
                "kitchenfloorFinishing": {
                  "type": "string"
                },
                "kitchenInteriorWallColor": {
                  "type": "string",
                  "pattern": "^0xFF[A-F0-9]{6}$"
                },
                "hasWeatherStation": {
                  "type": "boolean"
                },
                "entryStairs": {
                  "type": "boolean"
                },
                "hasFathometer": {
                  "type": "boolean"
                },
                "hasCleanWaterLevelIndicator": {
                  "type": "boolean"
                },
                "interiorBedroomWallColor": {
                  "type": "string",
                  "pattern": "^0xFF[A-F0-9]{6}$"
                },
                "deckFloorFinishMaterial": {
                  "type": "string"
                },
                "seaPodStatus": {
                  "type": "string"
                }
              },
              "example": {
                "SeaPodName": "Sea Breeze",
                "exteriorFinish": "exfinish",
                "exterioirColor": "0xff123456",
                "sparFinish": "sparFinish",
                "sparDesign": "sparDesign",
                "deckEnclosure": "deckEnclosure",
                "bedAndLivingRoomEnclousure": "bedAndLivingRoomEnclousure",
                "power": "power",
                "powerUtilities": [],
                "underWaterRoomFinishing": "underWaterRoomFinishing",
                "underWaterWindows": "underWaterWindows",
                "soundSystem": [],
                "masterBedroomFloorFinishing": [],
                "masterBedroomInteriorWallColor": "0xff123456",
                "livingRoomloorFinishing": "livingRoomloorFinishing",
                "livingRoomInteriorWallColor": "0xff123456",
                "kitchenfloorFinishing": "kitchenfloorFinishing",
                "kitchenInteriorWallColor": "0xff123456",
                "hasWeatherStation": false,
                "entryStairs": true,
                "hasFathometer": false,
                "hasCleanWaterLevelIndicator": false,
                "interiorBedroomWallColor": "0xff123456",
                "deckFloorFinishMaterial": "deckFloorFinishMaterial",
                "seaPodStatus": "seaPodStatus"
              }
            }
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Seapod data is required, Seapod Validation"
          },
          "404": {
            "description": "User Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
    "/seapods/{seapodId}/name": {
      "put": {
        "tags": [
          "Seapods"
        ],
        "summary": "Change Seapod Name",
        "description": "",
        "operationId": "changeSeapodName",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Seapod information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "seapodName": {
                  "type": "string"
                }
              },
              "example": {
                "seapodName": "Updated SeaPod Name"
              }
            }
          },
          {
            "in": "path",
            "name": "seapodId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the seapod"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid seapod id, Seapod name is required, Seapod name validation faild"
          },
          "403": {
            "description": "Access denied. you are not allowed to update the seapod name"
          },
          "404": {
            "description": "Seapod not found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
    "/seapods/{seapodId}/users/{userId}": {
      "delete": {
        "tags": [
          "Seapods"
        ],
        "summary": "Remove Member from Seapod",
        "description": "",
        "operationId": "removeUserFromSeapod",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "seapodId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the seapod"
          },
          {
            "in": "path",
            "name": "userId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the user to be removed"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid seapod id, Invalid user id, This member can't be removed,SeaPod Must have at least one member"
          },
          "403": {
            "description": "Access denied. you are not allowed to remove members"
          },
          "404": {
            "description": "Seapod not found, User not found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      },
      "get": {
        "tags": [
          "Seapods"
        ],
        "summary": "Get seaPodUser Details by Id",
        "description": "",
        "operationId": "getUserDetails",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "seapodId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the seapod"
          },
          {
            "in": "path",
            "name": "userId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the user"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid seapod id, Invalid user id"
          },
          "404": {
            "description": "Seapod not found, User not found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
...
    "/users/me/seapods": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User Seapods",
        "description": "",
        "operationId": "getUserSeapods",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Seapod Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
...
    "/permissions/{permissionId}/seapod/{seapodId}/name": {
      "put": {
        "tags": [
          "Permission"
        ],
        "summary": "Rename permission",
        "description": "",
        "operationId": "renamePermission",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Permission Information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "Name": {
                  "type": "string"
                }
              },
              "example": {
                "Name": "close friend"
              }
            }
          },
          {
            "in": "path",
            "name": "permissionId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the permission"
          },
          {
            "in": "path",
            "name": "seapodId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the seapod"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "permission name is required, Invalid seapod Id, Invalid permission Id, Duplicated Permission Name, Default Permission cannot be renamed"
          },
          "403": {
            "description": "Access denied. Not the Owner of the seapod"
          },
          "404": {
            "description": "Permission Not Found, Seapod Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
    "/permissions/{permissionId}/seapod/{seapodId}": {
      "delete": {
        "tags": [
          "Permission"
        ],
        "summary": "Delete permission",
        "description": "",
        "operationId": "deletePermission",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "permissionId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the permission"
          },
          {
            "in": "path",
            "name": "seapodId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the seapod"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid seapod Id, Invalid permission Id, Default Permission cannot be deleted"
          },
          "403": {
            "description": "Access denied. Not the Owner of the seapod"
          },
          "404": {
            "description": "Permission Not Found, Seapod Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
...
    "/permissions/{seapodId}/seapod": {
      "get": {
        "tags": [
          "Permission"
        ],
        "summary": "Get Seapod Permissions",
        "description": "",
        "operationId": "getSeapodPermissions",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "seapodId",
            "type": "string",
            "required": true,
            "description": "Numeric ID of the seapod"
          },
          {
            "in": "header",
            "name": "x-auth-token",
            "type": "string",
            "required": true
          },
          {
            "in": "header",
            "name": "hardwareId",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Invalid seapod Id"
          },
          "404": {
            "description": "Seapod Not Found"
          },
          "500": {
            "description": "Internal Server Error, Header Parameters is missing"
          }
        }
      }
    },
...
  }
}